# packages/api-client/openapi.json
{
  "components": {
    "schemas": {
...
      "CartItemAdd": {
        "additionalProperties": false,
        "properties": {
          "quantity": {
            "minimum": 1.0,
            "title": "Quantity",
            "type": "integer"
          },
          "sku_id": {
            "format": "uuid",
            "title": "Sku Id",
            "type": "string"
          }
        },
        "required": [
          "sku_id",
          "quantity"
        ],
        "title": "CartItemAdd",
        "type": "object"
      },
      "CartItemRead": {
        "additionalProperties": false,
        "properties": {
          "attributes": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Attributes",
            "type": "object"
          },
          "image_alt_he": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Alt He"
          },
          "image_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Url"
          },
          "is_active": {
            "title": "Is Active",
            "type": "boolean"
          },
          "line_total_agorot": {
            "title": "Line Total Agorot",
            "type": "integer"
          },
          "name_he": {
            "title": "Name He",
            "type": "string"
          },
          "product_id": {
            "format": "uuid",
            "title": "Product Id",
            "type": "string"
          },
          "product_type": {
            "title": "Product Type",
            "type": "string"
          },
          "quantity": {
            "title": "Quantity",
            "type": "integer"
          },
          "sku_code": {
            "title": "Sku Code",
            "type": "string"
          },
          "sku_id": {
            "format": "uuid",
            "title": "Sku Id",
            "type": "string"
          },
          "stock_quantity": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stock Quantity"
          },
          "unit_price_agorot": {
            "title": "Unit Price Agorot",
            "type": "integer"
          }
        },
        "required": [
          "sku_id",
          "sku_code",
          "product_id",
          "product_type",
          "name_he",
          "attributes",
          "quantity",
          "unit_price_agorot",
          "line_total_agorot",
          "stock_quantity",
          "is_active",
          "image_url",
          "image_alt_he"
        ],
        "title": "CartItemRead",
        "type": "object"
      },
      "CartItemSet": {
        "additionalProperties": false,
        "properties": {
          "quantity": {
            "minimum": 1.0,
            "title": "Quantity",
            "type": "integer"
          }
        },
        "required": [
          "quantity"
        ],
        "title": "CartItemSet",
        "type": "object"
      },
      "CartRead": {
        "additionalProperties": false,
        "properties": {
          "currency": {
            "const": "ILS",
            "title": "Currency",
            "type": "string"
          },
          "expires_at": {
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/CartItemRead"
            },
            "title": "Items",
            "type": "array"
          },
          "last_activity_at": {
            "format": "date-time",
            "title": "Last Activity At",
            "type": "string"
          },
          "shipping_agorot": {
            "title": "Shipping Agorot",
            "type": "integer"
          },
          "status": {
            "$ref": "#/components/schemas/CartStatus"
          },
          "subtotal_agorot": {
            "title": "Subtotal Agorot",
            "type": "integer"
          },
          "total_agorot": {
            "title": "Total Agorot",
            "type": "integer"
          },
          "total_quantity": {
            "title": "Total Quantity",
            "type": "integer"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "status",
          "items",
          "subtotal_agorot",
          "shipping_agorot",
          "total_agorot",
          "total_quantity",
          "currency",
          "last_activity_at",
          "expires_at",
          "version"
        ],
        "title": "CartRead",
        "type": "object"
      },
      "CartStatus": {
        "enum": [
          "active",
          "expired",
          "checked_out"
        ],
        "title": "CartStatus",
        "type": "string"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Coffix API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/cart": {
      "get": {
        "operationId": "get_cart_api_v1_cart_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CartRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "409": {
            "description": "Cart expired"
          }
        },
        "summary": "Get Cart",
        "tags": [
          "cart"
        ]
      }
    },
    "/api/v1/cart/items": {
      "post": {
        "operationId": "add_cart_item_api_v1_cart_items_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CartItemAdd"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CartRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "409": {
            "description": "Cart or stock conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Add Cart Item",
        "tags": [
          "cart"
        ]
      }
    },
    "/api/v1/cart/items/{sku_id}": {
      "delete": {
        "operationId": "delete_cart_item_api_v1_cart_items__sku_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "sku_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Sku Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CartRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "409": {
            "description": "Cart expired"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Cart Item",
        "tags": [
          "cart"
        ]
      },
      "put": {
        "operationId": "set_cart_item_api_v1_cart_items__sku_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "sku_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Sku Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CartItemSet"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CartRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "409": {
            "description": "Cart or stock conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Set Cart Item",
        "tags": [
          "cart"
        ]
      }
    },
...
  }
}